Gallery fix 2
